refactor SpawnpointType and SpawnpointTypes to its own package
