+++ src/main/java/io/github/minigamecore/api/spawnpoint/Spawnpoint.java
@@ -26,6 +26,7 @@
 package io.github.minigamecore.api.spawnpoint;
 
 import io.github.minigamecore.api.spawnpoint.immutable.ImmutableSpawnpoint;
+import io.github.minigamecore.api.spawnpoint.spawnpointtype.SpawnpointType;
 import io.github.minigamecore.api.util.manipulation.Mutable;
 import org.spongepowered.api.entity.Transform;
 import org.spongepowered.api.scoreboard.Team;
